Ensure continuous view scroll is smooth, hoist shared components into dedicated mock files
