+++ src/__tests__/components/SegmentView.test.tsx
@@ -22,12 +22,7 @@ jest.mock('../../components/AnalysisStore', () => ({
   useGlossDispatch: () => () => {},
 }));
 
-jest.mock('../../components/TokenChip', () => ({
-  __esModule: true,
-  MemoizedInertTokenChip({ token }: Readonly<{ token: Token }>) {
-    return <span>{token.surfaceText}</span>;
-  },
-}));
+jest.mock('../../components/TokenChip');
 
 jest.mock('../../components/PhraseBox', () => ({
   __esModule: true,
